+++ src/components/StatusBar/Status/index.tsx
@@ -28,7 +28,7 @@ const Status = () => {
       onMouseOver={handleOnMouseOver}
       onMouseLeave={handleOnMouseLeave}
       onClick={(evt) => evt.stopPropagation()}
-      className={`onboard_security_1 rounded-full flex items-center bg-status-red h-[26px] px-2.5 text-sm font-bold duration-500 ease-in-out text-black transition-all ${
+      className={`onboard_security_1 mr-1 rounded-[15px] flex items-center bg-status-red h-[32px] px-3.5 text-sm font-bold duration-500 ease-in-out text-black transition-all ${
         display ? 'scale-100 opacity-100' : 'scale-0 opacity-0'
       }`}
     >
@@ -38,7 +38,7 @@ const Status = () => {
         }`}
       >
         <div
-          className={`absolute top-1 left-3 transition-all delay-100 ${
+          className={`absolute top-1.5 left-4 transition-all delay-100 ${
             hover ? 'opacity-100' : 'delay-0 duration-0 opacity-0 h-[0px] w-[0px]'
           }`}
         >